pyproject.toml
[project]
name = "daily_philosopher"
version = "0.1.0"
description = "daily_philosopher using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "agentops>=0.3.21",
    "celery>=5.4.0",
    "crewai[tools]>=0.86.0,<1.0.0",
    "html2text>=2024.2.26",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "sendgrid>=6.11.0",
]

[project.scripts]
daily_philosopher = "daily_philosopher.main:run"
run_crew = "daily_philosopher.main:run"
train = "daily_philosopher.main:train"
replay = "daily_philosopher.main:replay"
test = "daily_philosopher.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
